% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiCorpus.R
\docType{package}
\name{ddicorpusR}
\alias{ddicorpusR}
\alias{ddiCorpus}
\alias{ddicorpusR-package}
\title{ddi CorpusR}
\format{
A tibble data with two variables: \code{drug_name}, \code{text}
}
\source{
\url{http://labda.inf.uc3m.es/DrugDDI/DrugDDI.html}
}
\usage{
ddiCorpus
}
\description{
A corpus annotated with drugdrug interaction.This dataset has been created by Hulat Group at University Carlos III of Madrid.
             Contact address: isegura@inf.uc3m.es or isegurabe@gmail.com
}
\keyword{datasets}
